requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
cryptography==2.2.2
docopt==0.6.2
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT==0.3.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
nose==1.3.7
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
pycparser==2.18
PyJWT==1.4.2
pytest==3.6.3
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
